--- a/package.json
+++ b/package.json
@@ -1,7 +1,7 @@
 {
   "name": "slow-json-stringify",
-  "version": "2.0.1",
-  "description": "The slowest JSON stringifier in the galaxy (:",
+  "version": "2.1.0",
+  "description": "The slow.. well actually fastest JSON stringifier in the galaxy.",
   "source": "src/sjs.mjs",
   "main": "dist/sjs.js",
   "module": "dist/sjs.mjs",