--- a/package.json
+++ b/package.json
@@ -50,5 +50,5 @@
     "url": "https://github.com/lucagez/slow-json-stringify/issues"
   },
   "homepage": "https://github.com/lucagez/slow-json-stringify#readme",
-  "sideeffects": false
+  "sideEffects": false
 }