--- a/src/_makeChunks.mjs
+++ b/src/_makeChunks.mjs
@@ -1,36 +1,47 @@
+const _stringRegex = /string/;
+
+const _replaceString = (type) =>
+  _stringRegex.test(type) ? '"__par__"' : "__par__";
+
+const _isLastRegex = /^("}|})/;
+
+// 3 possibilities after arbitrary property:
+// - ", => non-last string property
+// - , => non-last non-string property
+// - " => last string property
+const _matchStartRe = /^(\"\,|\,|\")/;
+
 /**
  * @param {string} str - prepared string already validated.
  * @param {array} queue - queue containing the property name to match
  * (used for building dynamic regex) needed for the preparation of
  * chunks used in different scenarios.
  */
-export default (str, queue) => str
-  // Matching prepared properties and replacing with target with or without
-  // double quotes.
-  // => Avoiding unnecessary concatenation of doublequotes during serialization.
-  .replace(/"\w+__sjs"/gm, type => (/string/.test(type) ? '"__par__"' : '__par__'))
-  .split('__par__')
-  .map((chunk, index, chunks) => {
+const _makeChunks = (str, queue) => {
+  const chunks = str
+      // Matching prepared properties and replacing with target with or without
+      // double quotes.
+      // => Avoiding unnecessary concatenation of doublequotes during serialization.
+      .replace(/"\w+__sjs"/gm, _replaceString)
+      .split("__par__"),
+    result = [];
+
+  for (let i = 0; i < chunks.length; ++i) {
+    const chunk = chunks[i];
+
     // Using dynamic regex to ensure that only the correct property
     // at the end of the string it's actually selected.
     // => e.g. ,"a":{"a": => ,"a":{
-    const matchProp = `("${(queue[index] || {}).name}":(\"?))$`;
-    const matchWhenLast = `(\,?)${matchProp}`;
+    const matchProp = `("${(queue[i] || {}).name}":(\"?))$`;
 
     // Check if current chunk is the last one inside a nested property
-    const isLast = /^("}|})/.test(chunks[index + 1] || '');
+    const isLast = _isLastRegex.test(chunks[i + 1] || "");
 
     // If the chunk is the last one the `isUndef` case should match
     // the preceding comma too.
-    const matchPropRe = new RegExp(isLast ? matchWhenLast : matchProp);
+    const matchPropRe = new RegExp(isLast ? `(\,?)${matchProp}` : matchProp);
 
-    // 3 possibilities after arbitrary property:
-    // - ", => non-last string property
-    // - , => non-last non-string property
-    // - " => last string property
-    const matchStartRe = /^(\"\,|\,|\")/;
-
-    return {
+    result.push({
       // notify that the chunk preceding the current one has not
       // its corresponding property undefined.
       // => This is a V8 optimization as it's way faster writing
@@ -38,12 +49,15 @@ export default (str, queue) => str
       flag: false,
       pure: chunk,
       // Without initial part
-      prevUndef: chunk.replace(matchStartRe, ''),
+      prevUndef: chunk.replace(_matchStartRe, ""),
       // Without property chars
-      isUndef: chunk.replace(matchPropRe, ''),
+      isUndef: chunk.replace(matchPropRe, ""),
       // Only remaining chars (can be zero chars)
-      bothUndef: chunk
-        .replace(matchStartRe, '')
-        .replace(matchPropRe, ''),
-    };
-  });
+      bothUndef: chunk.replace(_matchStartRe, "").replace(matchPropRe, ""),
+    });
+  }
+
+  return result;
+};
+
+export { _makeChunks };