--- a/src/_makeQueue.mjs
+++ b/src/_makeQueue.mjs
@@ -1,4 +1,30 @@
-import { _find } from './_utils';
+import { __find, _find } from "./_utils.mjs";
+
+const _sjsRegex = /__sjs/;
+
+function _prepareQueue(originalSchema, queue, obj, acc = []) {
+  if (_sjsRegex.test(obj)) {
+    const usedAcc = [...acc];
+    const find = __find(usedAcc);
+    const { serializer } = find(originalSchema);
+
+    queue.push({
+      serializer,
+      find,
+      name: acc[acc.length - 1],
+    });
+    return;
+  }
+
+  // Recursively going deeper.
+  // NOTE: While going deeper, the current prop is pushed into the accumulator
+  // to keep track of the position inside of the object.
+  const keys = Object.keys(obj);
+  for (let i = 0; i < keys.length; ++i) {
+    const key = keys[i];
+    _prepareQueue(originalSchema, queue, obj[key], [...acc, key]);
+  }
+}
 
 /**
  * @param {object} preparedSchema - schema already validated
@@ -6,32 +32,10 @@ import { _find } from './_utils';
  * @param {object} originalSchema - User provided schema
  * => contains array stringification serializers that are lost during preparation.
  */
-export default (preparedSchema, originalSchema) => {
+const _makeQueue = (preparedSchema, originalSchema) => {
   const queue = [];
-
-  // Defining a function inside an other function is slow.
-  // However it's OK for this use case as the queue creation is not time critical.
-  (function scoped(obj, acc = []) {
-    if (/__sjs/.test(obj)) {
-      const usedAcc = Array.from(acc);
-      const find = _find(usedAcc);
-      const { serializer } = find(originalSchema);
-
-      queue.push({
-        serializer,
-        find,
-        name: acc[acc.length - 1],
-      });
-      return;
-    }
-
-    // Recursively going deeper.
-    // NOTE: While going deeper, the current prop is pushed into the accumulator
-    // to keep track of the position inside of the object.
-    return Object
-      .keys(obj)
-      .map(prop => scoped(obj[prop], [...acc, prop]));
-  })(preparedSchema);
-
+  _prepareQueue(originalSchema, queue, preparedSchema);
   return queue;
 };
+
+export { _makeQueue };