--- a/src/_select.mjs
+++ b/src/_select.mjs
@@ -8,7 +8,7 @@
  * @param {any} value - value to serialize.
  * @param {number} index - position inside the queue.
  */
-const _select = (chunks) => (value, index) => {
+const _select = (chunks, lastIndex) => (value, index) => {
   const chunk = chunks[index];
 
   if (value !== undefined) {
@@ -20,12 +20,10 @@ const _select = (chunks) => (value, index) => {
 
   // If the current value is undefined set a flag on the next
   // chunk stating that the previous prop is undefined.
-  chunks[index + 1].flag = true;
+  if (index !== lastIndex)
+    chunks[index + 1].flag = true;
 
-  if (chunk.flag) {
-    return chunk.bothUndef;
-  }
-  return chunk.isUndef;
+  return chunk.flag ? chunk.bothUndef : chunk.isUndef;
 };
 
 export { _select };