--- a/src/_utils.mjs
+++ b/src/_utils.mjs
@@ -1,4 +1,3 @@
-
 /**
  * `_find` is a super fast deep property finder.
  * It dynamically build the function needed to reach the desired
@@ -11,16 +10,23 @@
  * @param {array} path - path to reach object property.
  */
 const _find = (path) => {
-  const { length } = path;
-
   let str = 'obj';
 
-  for (let i = 0; i < length; i++) {
-    str = str.replace(/^/, '(');
-    str += ` || {}).${path[i]}`;
+  for (let i = 0; i < path.length; ++i) {
+    str = `(${str}||{}).${path[i]}`;
   }
 
-  return eval(`((obj) => ${str})`);
+  return eval(`(obj=>${str})`);
+};
+
+const __find = (path) => {
+  let str = 'obj';
+
+  for (let i = 0; i < path.length; ++i) {
+    str += `?.['${path[i]}']`;
+  }
+
+  return eval(`(obj=>${str})`);
 };
 
 /**
@@ -35,13 +41,13 @@ const _makeArraySerializer = (serializer) => {
     return (array) => {
       // Stringifying more complex array using the provided sjs schema
       let acc = '';
-      const { length } = array;
-      for (let i = 0; i < length - 1; i++) {
+      const len = array.length - 1;
+      for (let i = 0; i < len; ++i) {
         acc += `${serializer(array[i])},`;
       }
 
       // Prevent slice for removing unnecessary comma.
-      acc += serializer(array[length - 1]);
+      acc += serializer(array[len]);
       return `[${acc}]`;
     };
   }
@@ -49,37 +55,40 @@ const _makeArraySerializer = (serializer) => {
   return array => JSON.stringify(array);
 };
 
-const TYPES = [
-  'number',
-  'string',
-  'boolean',
-  'array',
-  'null',
-];
+const TYPES = ['number', 'string', 'boolean', 'array', 'null'];
 
-const attr = (type, serializer) => {
-  if (!TYPES.includes(type)) {
-    throw new Error(`Expected one of: "number", "string", "boolean", "null". received "${type}" instead`);
+/*#__PURE__*/
+function checkType(type) {
+  if (typeof process !== 'undefined' && process.env.NODE_ENV !== 'production' && !TYPES.includes(type)) {
+    throw new Error(
+      `Expected one of: "number", "string", "boolean", "array", "null". received "${type}" instead`,
+    );
   }
+}
+
+/**
+ * @param type number|string|boolean|array|null
+ * @param serializer
+ * @returns
+ */
+const attr = (type, serializer) => {
+  /*#__PURE__*/checkType(type);
 
   const usedSerializer = serializer || (value => value);
 
   return {
     isSJS: true,
     type,
-    serializer: type === 'array'
-      ? _makeArraySerializer(serializer)
-      : usedSerializer,
+    serializer:
+      type === 'array' ? _makeArraySerializer(serializer) : usedSerializer,
   };
 };
 
 // Little utility for escaping convenience.
 // => if no regex is provided, a default one will be used.
-const defaultRegex = new RegExp('\\n|\\r|\\t|\\"|\\\\', 'gm');
-const escape = (regex = defaultRegex) => str => str.replace(regex, char => '\\' + char);
+const _defaultRegex = new RegExp('\\n|\\r|\\t|\\"|\\\\', 'gm');
+const _escapeCallback = char => '\\' + char;
+const escape = (regex = _defaultRegex) => str => str.replace(regex, _escapeCallback);
 
-export {
-  _find,
-  escape,
-  attr,
-};
+export { __find, _find, escape, attr };
+