--- a/src/_utils.mjs
+++ b/src/_utils.mjs
@@ -83,7 +83,7 @@ const attr = (type, serializer) => {
 
 // Little utility for escaping convenience.
 // => if no regex is provided, a default one will be used.
-const _defaultRegex = new RegExp('\\n|\\r|\\t|\\"|\\\\', 'gm');
+const _defaultRegex = /[\t\n\r"\\]/g
 const _escapeCallback = char => '\\' + char;
 const escape = (regex = _defaultRegex) => str => str.replace(regex, _escapeCallback);