--- a/src/sjs.mjs
+++ b/src/sjs.mjs
@@ -14,6 +14,7 @@ const sjs = (schema) => {
   // during schema preparation => e.g. array stringification method.
   const queue = _makeQueue(_preparedSchema, schema);
   const chunks = _makeChunks(_preparedString, queue);
+  const chunkLength = chunks.length - 1;
   const selectChunk = _select(chunks);
 
   // Exposed function
@@ -27,7 +28,7 @@ const sjs = (schema) => {
       temp += selectChunk(serializer(raw), i)
     }
 
-    const { flag, pure, prevUndef } = chunks[chunks.length - 1];
+    const { flag, pure, prevUndef } = chunks[chunkLength];
 
     return temp + (flag ? prevUndef : pure);
   };