@kurtextrem 2022-01-07
> Thank you! 🎉

There is not much more to optimize on the hot path as far as I can see.